Make a callback based ChatManager
